Modulariza o projeto
