+++ imagens/tog.ppm
@@ -1,0 +1,15 @@
+P3
+2 2
+255
+100
+150
+255
+29
+137
+0
+76
+149
+29
+53
+88
+22
