+++ modulos/constantes/constantes.h
@@ -1,0 +1,12 @@
+#ifndef _CONSTANTES_H_
+#define _CONSTANTES_H_
+
+#define SAIR 8
+
+#define LINHA_QUE_COMECAM_OS_PIXEIS 4
+
+#define RED 1
+#define GREEN 2
+#define BLUE 3
+
+#endif
