+++ modulos/funcoes/funcoes_gerais/funcoes_gerais.h
@@ -1,0 +1,12 @@
+#ifndef _FUN_GERAIS_H_
+#define _FUN_GERAIS_H_
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+FILE* pegaImagemPeloNome();
+void mostraMenu();
+void pegaOpcaoUsuario(int *opcao);
+
+#endif
